Refactoring: own file for transformation IBeXPattern -> Democles Pattern
